Rework of saving image
