+++ src/main/java/com/sda/eventapp/service/EventService.java
@@ -4,22 +4,15 @@ import com.sda.eventapp.dto.CommentView;
 import com.sda.eventapp.dto.EventView;
 import com.sda.eventapp.mapper.EventMapper;
 import com.sda.eventapp.model.Event;
-import com.sda.eventapp.model.Image;
 import com.sda.eventapp.model.User;
 import com.sda.eventapp.repository.EventRepository;
 import com.sda.eventapp.web.mvc.form.CreateCommentForm;
 import com.sda.eventapp.web.mvc.form.EventForm;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -29,8 +22,6 @@ import java.util.stream.StreamSupport;
 @Service
 @RequiredArgsConstructor
 public class EventService {
-    private final static String IMAGES_PATH = "src/main/resources/static/images/";
-
     private final EventRepository repository;
     private final CommentService commentService;
     private final ImageService imageService;
@@ -38,7 +29,7 @@ public class EventService {
 
     public Event save(EventForm form, User owner, MultipartFile file) {
         form.setOwner(owner);
-        form.setImage(solveImage(file));
+        form.setImage(imageService.solveImage(file));
         return repository.save(mapper.toEvent(form));
     }
 
@@ -49,7 +40,7 @@ public class EventService {
 
         if (file.getOriginalFilename() != null && !file.getOriginalFilename().isBlank()) {
             // if new file was uploaded
-            form.setImage(saveImageLocally(file));
+            form.setImage(imageService.solveImage(file));
         } else {
             // otherwise get image from db
             form.setImage(event.getImage());
@@ -161,80 +152,6 @@ public class EventService {
 
     public void saveComment(CreateCommentForm form, Long eventId, User loggedUser) {
         commentService.save(form, this.findById(eventId), loggedUser);
-    }
-
-    /**
-     * Taking {@link org.springframework.web.multipart.MultipartFile} and checks its name if it's {@link java.lang.String#isBlank()} or null.
-     * <br>
-     * If false, it does {@link com.sda.eventapp.service.EventService#saveImageLocally(MultipartFile)}
-     *
-     * @param file should be checked before if its of required file extension
-     * @return {@link com.sda.eventapp.model.Image} build from {@link org.springframework.web.multipart.MultipartFile}
-     */
-    private Image solveImage(MultipartFile file) {
-        Image image;
-        if (file.getOriginalFilename() == null || file.getOriginalFilename().isBlank()) {
-            image = getDefaultImage();
-        } else {
-            image = saveImageLocally(file);
-        }
-        return image;
-    }
-
-    private Image getDefaultImage() {
-        return imageService.buildDefaultImage(
-                Paths.get("").toAbsolutePath(),
-                IMAGES_PATH);
-    }
-
-    /**
-     * Creates directory in default location ->  {@link EventService#createImageDirectory()}.
-     * <br>
-     * Then builds image from file originalFilename -> {@link ImageService#buildImage(String
-     * fileName, Path absolutePath, String contentRootPath)}
-     * <br>
-     * Then checks if image of this name already exists in repository. If true, it will randomize
-     * its name suffix and will try again -> {@link EventService#randomizeFilename(String fileName)}
-     *
-     * @param file {@link org.springframework.web.multipart.MultipartFile} that is meant to be saved
-     *             locally and built {@link com.sda.eventapp.model.Image} out of it
-     * @return built and saved locally {@link com.sda.eventapp.model.Image}
-     */
-    private Image saveImageLocally(MultipartFile file) {
-        if (createImageDirectory()) {
-            log.debug("Directory created");
-        }
-        String originalFilename = file.getOriginalFilename();
-        Image image = imageService.buildImage(originalFilename, Paths.get("").toAbsolutePath(), IMAGES_PATH);
-
-        // TODO #004: This implementation of while loop is smelly. We should think about something else.
-        while (imageService.existsByFilename(image.getFilename())) {
-            image.setFilename(randomizeFilename(originalFilename));
-        }
-
-        Path fullPath = Paths.get(image.getPath() + image.getFilename());
-
-        try {
-            byte[] bytes = file.getBytes();
-            Files.write(fullPath, bytes);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        return image;
-    }
-
-    private String randomizeFilename(String filename) {
-        String salt = RandomStringUtils.random(10, true, false);
-        return salt + filename;
-    }
-
-    /**
-     * Creates directory named after path: {@link com.sda.eventapp.service.EventService#IMAGES_PATH} if it doesn't already exist.
-     *
-     * @return true in the same manner as {@link java.io.File#mkdirs()}
-     */
-    private boolean createImageDirectory() {
-        return new File(IMAGES_PATH).mkdirs();
     }
 
     public Event signUpForEvent(User user, Long eventId) {
